Fix relations and policies

+++ app/Http/Controllers/BookingController.php
@@ -5,18 +5,25 @@ namespace App\Http\Controllers;
 use App\Models\Booking;
 use App\Models\House;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Auth;
 
 class BookingController extends Controller
 {
     public function index()
     {
+        $this->authorize('viewAny', Booking::class);
+        if(Auth::user()->role == "admin")
+            $bookings = Booking::all()->sortByDesc('id');
+        else
+            $bookings = auth()->user()->bookings()->sortByDesc('id');
         return view('bookings.index', [
-            'bookings' => Booking::all()->sortByDesc('id')
+            'bookings' => $bookings
         ]);
     }
 
     public function create(House $house)
     {
+        $this->authorize('create', Booking::class);
         return view('bookings.create', [
             'house' => $house
         ]);
@@ -24,6 +31,7 @@ class BookingController extends Controller
 
     public function store(Request $request)
     {
+        $this->authorize('create', Booking::class);
         $data = $request->validate([
             'house_id' => ['integer', 'exists:houses,id', 'required'],
             'dateFrom' => ['date', 'after:today', 'required'],
@@ -46,6 +54,7 @@ class BookingController extends Controller
 
     public function show(Booking $booking)
     {
+        $this->authorize('view', $booking);
         return view('bookings.show', [
             'booking' => $booking
         ]);
@@ -53,22 +62,23 @@ class BookingController extends Controller
 
     public function edit(Booking $booking)
     {
+        $this->authorize('update', $booking);
         return view('bookings.edit', [
-            'booking' => $booking
+            'booking' => $booking,
+            'house' => $booking->house
         ]);
     }
 
     public function update(Request $request, Booking $booking)
     {
+        $this->authorize('update', $booking);
         $data = $request->validate([
             'house_id' => ['integer', 'exists:houses,id', 'required'],
             'dateFrom' => ['date', 'after:today', 'required'],
             'dateTo' => ['date', 'after:dateFrom', 'required'],
-            'user_id' => ['integer', 'exists:users,id', 'required']
         ]);
 
-        $data['dateFrom'] = date_format($data['dateFrom'], 'Y/m/d');
-        $data['dateTo'] = date_format($data['dateTo'], 'Y/m/d');
+        $data['user_id'] = auth()->user()->id;
 
         if(!$booking->update($data)){
             return redirect()->back()->withErrors([
@@ -84,13 +94,14 @@ class BookingController extends Controller
 
     public function destroy(Booking $booking)
     {
+        $this->authorize('update', $booking);
         if(!$booking->delete()){
             return redirect()->back()->withErrors([
                 'error' => 'Unable to delete booking'
             ]);
         }
         else{
-            return redirect()->route('booking.index')->with([
+            return redirect()->route('bookings.index')->with([
                 'status' => 'success'
             ]);
         }
